tests: move tests to corresponding modules
