+++ drive/src/drive/fee_pools/epochs/credit_distribution_pools.rs
@@ -127,54 +127,6 @@ mod tests {
     use crate::fee_pools::epochs::Epoch;
     use grovedb::Element;
 
-    mod update_epoch_storage_credits_for_distribution {
-
-        #[test]
-        fn test_error_if_epoch_pool_is_not_initiated() {
-            let drive = super::setup_drive_with_initial_state_structure();
-            let transaction = drive.grove.start_transaction();
-
-            let epoch = super::Epoch::new(7000);
-
-            let op = epoch.update_storage_credits_for_distribution_operation(42);
-
-            match drive.grove_apply_operation(op, false, Some(&transaction)) {
-                Ok(_) => assert!(
-                    false,
-                    "should not be able to update storage fee on uninit epochs pool"
-                ),
-                Err(e) => match e {
-                    super::error::Error::GroveDB(grovedb::Error::PathKeyNotFound(_)) => {
-                        assert!(true)
-                    }
-                    _ => assert!(false, "invalid error type"),
-                },
-            }
-        }
-
-        #[test]
-        fn test_value_is_set() {
-            let drive = super::setup_drive_with_initial_state_structure();
-            let transaction = drive.grove.start_transaction();
-
-            let epoch = super::Epoch::new(0);
-
-            let storage_fee = 42;
-
-            let op = epoch.update_storage_credits_for_distribution_operation(storage_fee);
-
-            drive
-                .grove_apply_operation(op, false, Some(&transaction))
-                .expect("should apply batch");
-
-            let stored_storage_fee = drive
-                .get_epoch_storage_credits_for_distribution(&epoch, Some(&transaction))
-                .expect("should get storage fee");
-
-            assert_eq!(stored_storage_fee, storage_fee);
-        }
-    }
-
     mod get_epoch_storage_credits_for_distribution {
         use crate::fee_pools::epochs_root_tree_key_constants::KEY_STORAGE_FEE_POOL;
 
@@ -226,53 +178,6 @@ mod tests {
                     _ => assert!(false, "invalid error type"),
                 },
             }
-        }
-    }
-
-    mod update_epoch_processing_credits_for_distribution {
-        #[test]
-        fn test_error_if_epoch_pool_is_not_initiated() {
-            let drive = super::setup_drive_with_initial_state_structure();
-            let transaction = drive.grove.start_transaction();
-
-            let epoch = super::Epoch::new(7000);
-
-            let op = epoch.update_processing_credits_for_distribution_operation(42);
-
-            match drive.grove_apply_operation(op, false, Some(&transaction)) {
-                Ok(_) => assert!(
-                    false,
-                    "should not be able to update processing fee on uninit epochs pool"
-                ),
-                Err(e) => match e {
-                    super::error::Error::GroveDB(grovedb::Error::PathKeyNotFound(_)) => {
-                        assert!(true)
-                    }
-                    _ => assert!(false, "invalid error type"),
-                },
-            }
-        }
-
-        #[test]
-        fn test_value_is_set() {
-            let drive = super::setup_drive_with_initial_state_structure();
-            let transaction = drive.grove.start_transaction();
-
-            let epoch = super::Epoch::new(0);
-
-            let processing_fee: u64 = 42;
-
-            let op = epoch.update_processing_credits_for_distribution_operation(42);
-
-            drive
-                .grove_apply_operation(op, false, Some(&transaction))
-                .expect("should apply batch");
-
-            let stored_processing_fee = drive
-                .get_epoch_processing_credits_for_distribution(&epoch, Some(&transaction))
-                .expect("should get processing fee");
-
-            assert_eq!(stored_processing_fee, processing_fee);
         }
     }
