+++ drive/src/fee_pools/epochs/mod.rs
@@ -22,6 +22,3 @@ impl Epoch {
         Self { index, key }
     }
 }
-
-// TODO Would be more convenient to have all methods here,
-//   because when you jump to implementation you arriving to this file and don't see any
