+++ drive/src/fee_pools/epochs/operations_factory.rs
@@ -445,6 +445,35 @@ mod tests {
         }
     }
 
+    mod update_proposer_block_count {
+        #[test]
+        fn test_value_is_set() {
+            let drive = super::setup_drive_with_initial_state_structure();
+            let transaction = drive.grove.start_transaction();
+
+            let pro_tx_hash: [u8; 32] = rand::random();
+            let block_count = 42;
+
+            let epoch = super::Epoch::new(0);
+
+            let mut batch = super::GroveDbOpBatch::new();
+
+            batch.push(epoch.init_proposers_tree_operation());
+
+            batch.push(epoch.update_proposer_block_count_operation(&pro_tx_hash, block_count));
+
+            drive
+                .grove_apply_batch(batch, false, Some(&transaction))
+                .expect("should apply batch");
+
+            let stored_block_count = drive
+                .get_epochs_proposer_block_count(&epoch, &pro_tx_hash, Some(&transaction))
+                .expect("should get proposer block count");
+
+            assert_eq!(stored_block_count, block_count);
+        }
+    }
+
     #[test]
     fn test_update_start_time() {
         let drive = setup_drive_with_initial_state_structure();
@@ -491,5 +520,230 @@ mod tests {
         assert_eq!(start_block_height, actual_start_block_height);
     }
 
-    // TODO: Find and move all related tests here
+    mod update_epoch_processing_credits_for_distribution {
+        use crate::error;
+
+        #[test]
+        fn test_error_if_epoch_pool_is_not_initiated() {
+            let drive = super::setup_drive_with_initial_state_structure();
+            let transaction = drive.grove.start_transaction();
+
+            let epoch = super::Epoch::new(7000);
+
+            let op = epoch.update_processing_credits_for_distribution_operation(42);
+
+            match drive.grove_apply_operation(op, false, Some(&transaction)) {
+                Ok(_) => assert!(
+                    false,
+                    "should not be able to update processing fee on uninit epochs pool"
+                ),
+                Err(e) => match e {
+                    error::Error::GroveDB(grovedb::Error::PathKeyNotFound(_)) => {
+                        assert!(true)
+                    }
+                    _ => assert!(false, "invalid error type"),
+                },
+            }
+        }
+
+        #[test]
+        fn test_value_is_set() {
+            let drive = super::setup_drive_with_initial_state_structure();
+            let transaction = drive.grove.start_transaction();
+
+            let epoch = super::Epoch::new(0);
+
+            let processing_fee: u64 = 42;
+
+            let op = epoch.update_processing_credits_for_distribution_operation(42);
+
+            drive
+                .grove_apply_operation(op, false, Some(&transaction))
+                .expect("should apply batch");
+
+            let stored_processing_fee = drive
+                .get_epoch_processing_credits_for_distribution(&epoch, Some(&transaction))
+                .expect("should get processing fee");
+
+            assert_eq!(stored_processing_fee, processing_fee);
+        }
+    }
+
+    mod update_epoch_storage_credits_for_distribution {
+        use crate::error;
+
+        #[test]
+        fn test_error_if_epoch_pool_is_not_initiated() {
+            let drive = super::setup_drive_with_initial_state_structure();
+            let transaction = drive.grove.start_transaction();
+
+            let epoch = super::Epoch::new(7000);
+
+            let op = epoch.update_storage_credits_for_distribution_operation(42);
+
+            match drive.grove_apply_operation(op, false, Some(&transaction)) {
+                Ok(_) => assert!(
+                    false,
+                    "should not be able to update storage fee on uninit epochs pool"
+                ),
+                Err(e) => match e {
+                    error::Error::GroveDB(grovedb::Error::PathKeyNotFound(_)) => {
+                        assert!(true)
+                    }
+                    _ => assert!(false, "invalid error type"),
+                },
+            }
+        }
+
+        #[test]
+        fn test_value_is_set() {
+            let drive = super::setup_drive_with_initial_state_structure();
+            let transaction = drive.grove.start_transaction();
+
+            let epoch = super::Epoch::new(0);
+
+            let storage_fee = 42;
+
+            let op = epoch.update_storage_credits_for_distribution_operation(storage_fee);
+
+            drive
+                .grove_apply_operation(op, false, Some(&transaction))
+                .expect("should apply batch");
+
+            let stored_storage_fee = drive
+                .get_epoch_storage_credits_for_distribution(&epoch, Some(&transaction))
+                .expect("should get storage fee");
+
+            assert_eq!(stored_storage_fee, storage_fee);
+        }
+    }
+
+    mod delete_proposers_tree {
+        use crate::fee_pools::epochs::epoch_key_constants::KEY_PROPOSERS;
+
+        #[test]
+        fn test_values_has_been_deleted() {
+            let drive = super::setup_drive_with_initial_state_structure();
+            let transaction = drive.grove.start_transaction();
+
+            let epoch = super::Epoch::new(0);
+
+            let mut batch = super::GroveDbOpBatch::new();
+
+            batch.push(epoch.init_proposers_tree_operation());
+
+            // Apply proposers tree
+            drive
+                .grove_apply_batch(batch, false, Some(&transaction))
+                .expect("should apply batch");
+
+            let mut batch = super::GroveDbOpBatch::new();
+
+            batch.push(epoch.delete_proposers_tree_operation());
+
+            drive
+                .grove_apply_batch(batch, false, Some(&transaction))
+                .expect("should apply batch");
+
+            match drive
+                .grove
+                .get(
+                    epoch.get_path(),
+                    KEY_PROPOSERS.as_slice(),
+                    Some(&transaction),
+                )
+                .unwrap()
+            {
+                Ok(_) => assert!(false, "expect tree not exists"),
+                Err(e) => match e {
+                    grovedb::Error::PathKeyNotFound(_) => assert!(true),
+                    _ => assert!(false, "invalid error type"),
+                },
+            }
+        }
+    }
+
+    mod delete_proposers {
+        #[test]
+        fn test_values_are_being_deleted() {
+            let drive = super::setup_drive_with_initial_state_structure();
+            let transaction = drive.grove.start_transaction();
+
+            let epoch = super::Epoch::new(0);
+
+            let mut batch = super::GroveDbOpBatch::new();
+
+            batch.push(epoch.init_proposers_tree_operation());
+
+            // Apply proposers tree
+            drive
+                .grove_apply_batch(batch, false, Some(&transaction))
+                .expect("should apply batch");
+
+            let pro_tx_hashes: Vec<[u8; 32]> = (0..10).map(|_| rand::random()).collect();
+
+            let mut batch = super::GroveDbOpBatch::new();
+
+            for pro_tx_hash in pro_tx_hashes.iter() {
+                batch.push(epoch.update_proposer_block_count_operation(pro_tx_hash, 1));
+            }
+
+            // Apply proposers block count updates
+            drive
+                .grove_apply_batch(batch, false, Some(&transaction))
+                .expect("should apply batch");
+
+            let mut stored_proposers = drive
+                .get_epoch_proposers(&epoch, 20, Some(&transaction))
+                .expect("should get proposers");
+
+            let mut awaited_result = pro_tx_hashes
+                .iter()
+                .map(|hash| (hash.to_vec(), 1))
+                .collect::<Vec<(Vec<u8>, u64)>>();
+
+            // sort both result to be able to compare them
+            stored_proposers.sort();
+            awaited_result.sort();
+
+            assert_eq!(stored_proposers, awaited_result);
+
+            let deleted_pro_tx_hashes = vec![
+                awaited_result.get(0).unwrap().0.clone(),
+                awaited_result.get(1).unwrap().0.clone(),
+            ];
+
+            // remove items we deleted
+            awaited_result.remove(0);
+            awaited_result.remove(1);
+
+            let mut batch = super::GroveDbOpBatch::new();
+
+            epoch.add_delete_proposers_operations(deleted_pro_tx_hashes, &mut batch);
+
+            // Apply proposers deletion
+            drive
+                .grove_apply_batch(batch, false, Some(&transaction))
+                .expect("should apply batch");
+
+            let stored_proposers = drive
+                .get_epoch_proposers(&epoch, 20, Some(&transaction))
+                .expect("should get proposers");
+
+            let mut stored_hexes: Vec<String> = stored_proposers
+                .iter()
+                .map(|(hash, _)| hex::encode(hash))
+                .collect();
+
+            let mut awaited_hexes: Vec<String> = stored_proposers
+                .iter()
+                .map(|(hash, _)| hex::encode(hash))
+                .collect();
+
+            stored_hexes.sort();
+            awaited_hexes.sort();
+
+            assert_eq!(stored_hexes, awaited_hexes);
+        }
+    }
 }
